Add doom emacs configs
I tried updating spacemacs to the develop branch,
as I was seeing a bunch of errors repeatedly displaying
related to org projectile and other packages,
all of which are just due to out of date packages.
The develop branch is years (!) ahead of master,
so it fixes a lot of those things.
But, after diffing and cleaning my dotspacemacs file
and trying to get back into regular use,
it seemed like things were just unbearably sluggish.
Helm was slow, there seemed to be random hangs
doing ordinary things, and it no longer felt nice.
I also was trying to adjust my modeline,
and found that vim-powerline looked amazing,
but didn't seem to be able to handle narrow windows.

So, I got fed up and I'm trying out doom emacs,
which at first blush is incredibly fast and pleasant!